+++ tests/tear_down.sh
@@ -1,0 +1,2 @@
+sudo ip link del vtesteth0
+sudo ip link del vtesteth1